refactor: split utils and core
